refactor(whiteboard): Reorg component hierarchy, clean up extraneous hook calls
