+++ demo-apps/whiteboard-basic/src/utils/helpers.js
@@ -1,0 +1,27 @@
+const rectangleColors = [
+  '#FF5733',
+  '#FFC300',
+  '#FFDC00',
+  '#D2FF00',
+  '#66FF00',
+  '#00FF66',
+  '#00FFD2',
+  '#0095FF',
+  '#0044FF',
+  '#002BFF',
+  '#5600FF',
+  '#B200FF',
+  '#FF00E6',
+  '#FF0052',
+  '#FF0033',
+];
+
+function getRandomInt(max) {
+  return Math.floor(Math.random() * max);
+}
+
+function getRandomColor() {
+  return rectangleColors[getRandomInt(rectangleColors.length)];
+}
+
+export { getRandomInt, getRandomColor };
